Struct st7735_lcd::ST7735

pub struct ST7735<SPI, DC, RST> where
    SPI: Write<u8>,
    DC: OutputPin,
    RST: OutputPin,  {
    _spi: SPI,
    _dc: DC,
    rst: RST,
    rgb: bool,
    inverted: bool,
    dx: u16,
    dy: u16,
}

ST7735 driver to connect to TFT displays.

Fields
_spi: SPI

SPI

_dc: DC

Data/command pin.

rst: RST

Reset pin.

rgb: bool

Whether the display is RGB (true) or BGR (false)

inverted: bool

Whether the colours are inverted (true) or not (false)

dx: u16

Global image offset

dy: u16
Methods

impl<SPI, DC, RST> ST7735<SPI, DC, RST> where
SPI: Write<u8>,
DC: OutputPin,
RST: OutputPin,

pub fn new(spi: SPI, dc: DC, rst: RST, rgb: bool, inverted: bool) -> Self

Creates a new driver instance that uses hardware SPI.

pub fn init<DELAY>(&mut self, delay: &mut DELAY) -> Result<(), ()> where
DELAY: DelayMs<u8>,

Runs commands to initialize the display.

pub fn hard_reset(&mut self) -> Result<(), ()>

fn write_command(
&mut self,
command: Instruction,
params: Option<&[u8]>
) -> Result<(), ()>

fn write_data(&mut self, data: &[u8]) -> Result<(), ()>

fn write_word(&mut self, value: u16) -> Result<(), ()>

Writes a data word to the display.

pub fn set_orientation(&mut self, orientation: &Orientation) -> Result<(), ()>

pub fn set_offset(&mut self, dx: u16, dy: u16)

Sets the global offset of the displayed image

fn set_address_window(
&mut self,
sx: u16,
sy: u16,
ex: u16,
ey: u16
) -> Result<(), ()>

Sets the address window for the display.

pub fn set_pixel(&mut self, x: u16, y: u16, color: u16) -> Result<(), ()>

Sets a pixel color at the given coords.

pub fn write_pixels<P: IntoIterator<Item = u16>>(
&mut self,
colors: P
) -> Result<(), ()>

Writes pixel colors sequentially into the current drawing window

pub fn set_pixels<P: IntoIterator<Item = u16>>(
&mut self,
sx: u16,
sy: u16,
ex: u16,
ey: u16,
colors: P
) -> Result<(), ()>

Sets pixel colors at the given drawing window